implemented: hadoop jar hdsf.jar put <src> <rst>

+++ hdfs/src/streams.java
@@ -1,11 +1,12 @@
 package tools;
 
-import java.io.File;
+//import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
-import java.io.OutputStream;
+//import java.io.OutputStream;
 import java.io.FileInputStream;
-import java.io.FileOutputStream;
+//import java.io.FileOutputStream;
+
 import java.lang.IllegalArgumentException;
 
 import org.apache.hadoop.conf.Configured;
@@ -13,8 +14,11 @@ import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
 import org.apache.hadoop.io.IOUtils;
-import org.apache.hadoop.util.Tool;
-import org.apache.hadoop.util.ToolRunner;
+//import org.apache.hadoop.util.Tool;
+//import org.apache.hadoop.util.ToolRunner;
+
+import org.apache.hadoop.fs.FSDataInputStream;
+import org.apache.hadoop.fs.FSDataOutputStream;
 
 public class streams {
 
@@ -23,38 +27,42 @@ public class streams {
       throw new IllegalArgumentException("argument is null.");
     }
   }
-  
-  private static boolean isLocal(String filename) {
-    File local = new File(filename);
-    return (local.exists() && ! local.isDirectory());
-  }
 
-  private static InputStream istream(String filename, FileSystem fs)
-    throws IllegalArgumentException, IOException {
-	if (isLocal(filename)) {
+
+
+/*
+  private static InputStream istream(String filename, FileSystem fs) throws IllegalArgumentException, IOException {
+        File local = new File(filename);
+	if (local.exists() && ! local.isDirectory()) {
 	    return new FileInputStream(filename);
 	}
-	Path hdfs = new Path(filename);
-        return fs.open(hdfs);
+        return fs.open(new Path(filename));
   }
 
-  private static OutputStream ostream(String filename)
-    throws IllegalArgumentException, IOException {
+  private static OutputStream ostream(String filename, FileSystem fs) throws IllegalArgumentException, IOException {
+        File local = new File(filename);
+	try {
+	
+	}
+
+	if (!local.exists() && local.isDirectory()) {
+	    return new FileInputStream(filename);
+	}
+
     AssertNotNull(filename);
     return new FileOutputStream(filename);
   }
+*/
 
-  public static void cp(String src_uri, String dst_uri)
-    throws IllegalArgumentException, IOException {
-        AssertNotNull(src_uri);
-	AssertNotNull(dst_uri);
-    	Configuration conf = new Configuration();
-	FileSystem fs = FileSystem.get(conf);
-	InputStream istream = istream(src_uri, fs);
-	AssertNotNull(istream);
-	OutputStream ostream = ostream(dst_uri);
-	AssertNotNull(ostream);
-
-	IOUtils.copyBytes(istream, ostream, conf, true);
+    public static void put(String ifile, String ofile) throws IllegalArgumentException, IOException {
+	AssertNotNull(ifile);
+	AssertNotNull(ofile);
+	Configuration conf = new Configuration();
+	FileSystem.get(conf).copyFromLocalFile(new Path(ifile), new Path(ofile));
+/*	
+	InputStream is = new FileInputStream(ifile);
+	FSDataOutputStream os = FileSystem.get(conf).create(new Path(ofile));
+	IOUtils.copyBytes(is, os, conf, true);
+*/
     }
 }
